tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-play-state_inherit.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-play-state: inherit</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-play-state-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="The computed value of 'animation-play-state' is
   the same as the specified value">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move {
         from {
            left: 0px;
         }
         to {
            left: 150px;
         }
      }
      div {
         width: 50px;
         height: 100px;
         background: gray;
         position: relative;
         animation-duration: 1s;
         animation-timing-function: steps(3);
         animation-fill-mode: forwards;
      }
      body > div {
         border-right: 50px solid red;
      }
      div > div {
         width: 100px;
         background: blue;
         animation-play-state: inherit;
      }
      .test, .test > div {
         animation-name: move;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var divs = document.querySelectorAll('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var outerStyle = getComputedStyle(divs[0], null);
         var innerStyle = getComputedStyle(divs[1], null);
         var passed = false;
         setTimeout(function() {
            divs[0].style.setProperty('animation-play-state', 'paused', '');
            passed = (outerStyle.left == '50px' && innerStyle.left == '50px');
         }, 500);
         setTimeout(function() {
            var state = innerStyle.getPropertyValue('animation-play-state');
            passed = (passed && state == 'paused' && outerStyle.left == '50px'
                      && innerStyle.left == '50px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         }, 1000);
         divs[0].className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue block below should move rightwards once, revealing a gray
   rectangle but no red.</p>
   <p>Testing...</p>
   <div>
      <div></div>
   </div>
</body>
</html>
